+++ src/webviews/copilotOnRails/views/styles/createProjectView.scss
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+@use './disabledButton';
+
 .createProjectView {
     display: flex;
     flex-direction: column;
