+++ src/webviews/copilotOnRails/views/styles/deployResultView.scss
@@ -3,6 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+@use './disabledButton';
 @use './fluentTooltipPortal';
 
 .deployResultView {
