+++ src/webviews/copilotOnRails/views/styles/frontendPreviewView.scss
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+@use './disabledButton';
+
 .frontendPreviewView {
     display: flex;
     flex-direction: column;
