+++ src/webviews/copilotOnRails/views/styles/loadingView.scss
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+@use './disabledButton';
+
 .loadingView {
     display: flex;
     flex-direction: column;
