+++ src/webviews/copilotOnRails/views/styles/nextStepsView.scss
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+@use './disabledButton';
+
 .nextStepsView {
     display: flex;
     flex-direction: column;
