+++ src/webviews/copilotOnRails/views/styles/scaffoldPlanView.scss
@@ -3,7 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+@use './disabledButton';
 @use './fluentTooltipPortal';
+@use './installedChip' as installedChipStyles;
 @import './uiPreviewCard.scss';
 
 .fui-DialogSurface {
@@ -96,14 +98,6 @@
 
             .fui-Button {
                 overflow: visible;
-            }
-
-            .fui-Button[data-appearance='primary']:disabled,
-            .fui-Button--primary:disabled {
-                background-color: var(--vscode-button-background) !important;
-                color: var(--vscode-button-foreground) !important;
-                border-color: var(--vscode-button-background) !important;
-                opacity: 0.6;
             }
 
             .feedbackIconWrapper {
@@ -437,29 +431,7 @@
             }
         }
 
-        .installedChip {
-            display: inline-flex;
-            align-items: center;
-            gap: 4px;
-            white-space: nowrap;
-            font-size: 0.85em;
-
-            .codicon {
-                font-size: 13px;
-            }
-
-            &.installed-installed {
-                color: var(--vscode-testing-iconPassed, var(--vscode-charts-green));
-            }
-
-            &.installed-missing {
-                color: var(--vscode-testing-iconFailed, var(--vscode-errorForeground));
-            }
-
-            &.installed-unknown {
-                color: var(--vscode-descriptionForeground);
-            }
-        }
+        @include installedChipStyles.installedChip;
 
         .installLink {
             display: inline-flex;
